Fix: prefix switch for receive QR (#2795)
